// src/types/empWorkType.ts
class EmployeeWorkData {
    employeeId: number;
    email: string;
    password: string;
    // permissionId: string;
    education: string;
    experiance: number;
    department: string;
    role: string;
    designation: string;
    reporting_lead: string;
    package: number;
    joining: Date;
}
  
export default EmployeeWorkData;